% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_map.R
\name{add_to_map}
\alias{add_to_map}
\title{Add a layer to your map, regarding a particular type of OSM key-value-geometry object.}
\usage{
add_to_map(basemap, gosmdata, layergroup = "additional", color = "red")
}
\arguments{
\item{basemap}{the basemap which you wish to add elements to}

\item{gosmdata}{the gosmdata element you wish to add to the map, as returned by pick_osmdata()}

\item{layergroup}{in which layer grouping elements should this gosmdata element be displayed (defaults to "additional")}

\item{color}{color of the data items}
}
\value{
a leaflet map
}
\description{
Add a layer to your map, regarding a particular type of OSM key-value-geometry object.
}
\examples{
g_bridges=get_osmdata("man_made","bridge","confluence") \%>\%
  pick_osmdata(geom="osm_polygons") \%>\%
  detail_element()
g_parks=get_osmdata("leisure","park","confluence") \%>\%
  pick_osmdata(geom="osm_polygons") \%>\%
  detail_element()
map=base_map("confluence") \%>\%
  add_to_map(g_bridges) \%>\%
  add_to_map(g_parks, color="green")
map

g_vineyard=get_osmdata("landuse","vineyard","condrieu") \%>\%
  pick_osmdata(geom="osm_polygons") \%>\%
  detail_element()
map=base_map("condrieu") \%>\%
  add_to_map(g_vineyard, color="purple")

}
