% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_table.R
\name{format_table}
\alias{format_table}
\title{Format a table for better display on shiny interface}
\usage{
format_table(
  tib,
  large = c("text_en", "text", "context"),
  medium = c("snippet", "trans_snippet")
)
}
\arguments{
\item{tib}{the table to format}

\item{large}{the names of columns to display as large (600px) columns. Defaults to c("text_en","text","context").}

\item{medium}{the names of columns to display as medium (300px) columns. Defaults to c("snippet","trans_snippet").}
}
\description{
Format a table for better display on shiny interface
}
