% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_query.R
\name{process_query}
\alias{process_query}
\title{Format a safe SQL query}
\usage{
process_query(search_string, column)
}
\arguments{
\item{search_string}{the query search condition}

\item{column}{the column in which the words are searched}
}
\description{
Format a safe SQL query
}
